Refactor rounding mixin to class-based mixin for numbers
Replaces the object-based RoundingMixin with a class-based mixin (WithRounding) for better TypeScript compatibility and cleaner inheritance. Updates the SimpleNumber component to use the new mixin pattern, improving maintainability and type safety.
