+++ components/src/components/Common/Rounding.mixin.ts
@@ -17,101 +17,6 @@ export interface RoundingConfig {
     decimalPlaces: number;
 }
 
-export const RoundingMixin = {
-    /**
-     * Apply rounding to a numeric value based on component configuration
-     */
-    applyRounding(value: number): number {
-        const roundingConfig = this.component.rounding as RoundingConfig;
-
-        if (!roundingConfig?.enabled) {
-            return value;
-        }
-        
-        const decimalPlaces = roundingConfig.decimalPlaces || 2;
-        const multiplier = Math.pow(10, decimalPlaces);
-        
-        switch (roundingConfig.method) {
-            case 'floor':
-                return Math.floor(value * multiplier) / multiplier;
-            case 'ceil':
-                return Math.ceil(value * multiplier) / multiplier;
-            case 'round':
-            default:
-                return Math.round(value * multiplier) / multiplier;
-        }
-    },
-
-    /**
-     * Process value with rounding if applicable
-     */
-    processValueWithRounding(value: any): any {
-        if (value === null || value === undefined || value === '') {
-            return value;
-        }
-        
-        const numValue = parseFloat(value);
-        if (isNaN(numValue)) {
-            return value;
-        }
-        
-        return this.applyRounding(numValue);
-    },
-
-    /**
-     * Override setValue to apply rounding
-     */
-    setValue(value: any, flags: any = {}) {
-        const processedValue = this.processValueWithRounding(value);
-        if (!this._originalSetValue) {
-            this._originalSetValue = Object.getPrototypeOf(Object.getPrototypeOf(this)).setValue;
-        }
-        return this._originalSetValue.call(this, processedValue, flags);
-    },
-
-    /**
-     * Override getValue to apply rounding
-     */
-    getValue() {
-        if (!this._originalGetValue) {
-            this._originalGetValue = Object.getPrototypeOf(Object.getPrototypeOf(this)).getValue;
-        }
-        const value = this._originalGetValue.call(this);
-        return this.processValueWithRounding(value);
-    },
-
-    /**
-     * Override calculateValue to apply rounding
-     */
-    calculateValue(data: any, flags: any, row: any) {
-        if (!this._originalCalculateValue) {
-            this._originalCalculateValue = Object.getPrototypeOf(Object.getPrototypeOf(this)).calculateValue;
-        }
-        const value = this._originalCalculateValue.call(this, data, flags, row);
-        return this.processValueWithRounding(value);
-    },
-
-    /**
-     * Override getValueAsString for consistent display
-     */
-    getValueAsString(value: any, options: any) {
-        const roundingConfig = this.component.rounding as RoundingConfig;
-        
-        if (roundingConfig && roundingConfig.enabled && value !== null && value !== undefined && value !== '') {
-            const numValue = parseFloat(value);
-            if (!isNaN(numValue)) {
-                const roundedValue = this.applyRounding(numValue);
-                return roundedValue.toFixed(roundingConfig.decimalPlaces || 2);
-            }
-        }
-        
-        if (!this._originalGetValueAsString) {
-            this._originalGetValueAsString = Object.getPrototypeOf(Object.getPrototypeOf(this)).getValueAsString;
-        }
-        return this._originalGetValueAsString.call(this, value, options);
-    }
-};
-
 // Add this to your Rounding.mixin.ts file
 export const RoundingEditFormComponents = [
     {
